pyproject.toml
[project]
name = "image-framer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.13",
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "pillow>=11.3.0",
    "pygments>=2.19.2",
    "python-dotenv>=1.1.1",
]
